Refactored the createFulfilled and createRejected methods
### Description
This change improves the `Swoole` class in the `Appwrite\GraphQL\Promises\Adapter` namespace by adding try-catch blocks to the `create` method to catch any exception thrown by the resolver function. This way, it will be rejected and passed to the $reject function. Additionally, I also refactored the `createFulfilled` and `createRejected` methods to use the `create` method and pass the appropriate callbacks. I also added some inline comments for better understanding of the code.

### Changes
- Add try-catch block in the `create` method to catch any exception thrown by the resolver function.
- Refactor the `createFulfilled` and `createRejected` methods to use the `create` method and pass the appropriate callbacks.
- Add inline comments for better understanding of the code.