Add a secrets attribute to the identities collection
These secrets can be used to store data from the provider that may or
may not be sensitive.

For example, this will be used by the migration API when connecting to
Firebase to store the service account used for the migration.

This data will only be used internally inside Appwrite and not exposed
to an end user or developer.