Fix _APP_EXECUTOR_HOST for upgrades
As of 1.4.0, the hostname of the executor should be executor rather than
appwrite-executor. The problem is our install command always uses the
existing installation's variable values as the default for the upgrade.
This means the _APP_EXECUTOR_HOST will retain it's old value.

This PR Adds an overwrite key to variables.php to allow overwriting the
variable value regardless of whatever was in the previous installation.