Fix delete function deployment
Prior to this, deleting a VCS deployment would fail because the
deployment path for VCS deployments are empty. Because an error gets
thrown, the delete function would stop and not delete the build files.

This change updates the worker to only delete the storage files if the
path is set.