Fix avatars tests
Because the tests referenced images on the appwrite.io website, but the
appwrite.io website changed, the tests broke. This updates the tests to
reference valid images/URLs.

In addition, the favicon test was changed to use github.com because
appwrite.io has an SVG favicon which the avatars API does not support.