fix(storage): add missing chunkId param to create file abuse key
A previous PR added chunkId as an abuse key param to fix a problem where uploading multiple chunks of a file would trigger the `general_rate_limit_exceeded` error. For some reason, the parameter disappeared from the abuse key causing a regression. This PR adds the param back to fix the regression.

Reference:
* https://github.com/appwrite/appwrite/issues/7879
* https://github.com/appwrite/appwrite/pull/7154