docker-compose: mariadb - innodb-flush-method deprecated
From https://mariadb.com/kb/en/innodb-system-variables/#innodb_flush_method
the innodb_flush_method is deprecated and is being replaced with
4 other variables affecting the write behaviour.

Its unclear why this was originally needed. The O_DIRECT default change
in 10.6 would have a negative impact if the innodb_buffer_pool was
insufficient to handle the full active dataset. If this is the case
--loose-innodb_data_file_buffering=OFF would be a compatible way to
support this.

As this variable is deprecated, we add the syntax to avoid failure
on start, even though its removal is going to be many years from now.