fix: Disable validation for project DBs during migration
Skip the validation so that queries that typically fail can pass. This
is fine because the validation is failing on internal attributes such
as an $id for cache key with * and $id for attributes that are too long
because we prefix the $id with the internal database and collection id.