fix(health): fix flaky certificate test
There seems to be a load balancer in front of www.google.com that has
either a certificate issuer of "Google Trust Services LLC" or
"Google Trust Services". This causes the test to fail intermittently.

This PR updates the test to check for the substring
"Google Trust Services" instead of the exact string
"Google Trust Services LLC".