fix(users): ensure user can delete session
The session document created by users.createSession() was missing
delete permissions for the user so when the user tried to delete it,
they got a 401 error.

This PR ensure the permissions are added just like if the document was
created from the Account API so that the user has access to delete the
document.