fix: lookup identity by email in OAuth2 login flow
Previously, we were only checking for existing identities by provider user ID during OAuth2 login. This could lead to an error if someone has linked a different email address to their account already, but log in using a different OAuth2 provider that shares the same email.

We initially required the user to connect the OAuth2 provider before using it to prevent a bad actor from hijacking an account by creating an account on the OAuth2 provider with the same email. However, this created a poor user experience. In addition, this doesn't stop the same thing from happening if a bad actor creates an account in the other OAuth2 provider with the same email as the Appwrite account email.